requirements.txt
flask==2.3.3
numpy==1.24.4
opencv-python==4.9.0.80
pyttsx3==2.90
transformers==4.38.2
torch==2.6.0
Pillow==10.3.0
git+https://github.com/microsoft/GenerativeImage2Text.git
deep-translator==1.11.4
SpeechRecognition==3.10.0
pyaudio==0.2.14

